pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "Meta-Roles"
version = "0.1.3"
description = "A discord bot made with naff."
authors = ["Maxence Lagalle <contact@maxence-lagalle.fr>"]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.20.0"
orjson = "^3.6.7"
interactions.py = "^5.1.0"
tortoise-orm = "^0.19.3"
aerich = "^0.7.1"
sentry-sdk = "^1.23.1"

[tool.poetry.dev-dependencies]


[tool.aerich]
tortoise_orm = "core.init_tortoise.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
